refactor: Relocate knowledge_stats from orbit-core to orbit-knowledge::… [ORB-00214]
Relocate knowledge_stats from orbit-core to orbit-knowledge::metrics

Planned-By: codex
