+++ .orbit/learnings/L-0027/learning.yaml
@@ -1,0 +1,20 @@
+schema_version: 1
+id: L-0027
+status: active
+scope:
+  paths:
+  - crates/*/Cargo.toml
+  - ARCHITECTURE.md
+  - scripts/check-dependency-direction.sh
+  tags:
+  - rust
+  - dependencies
+summary: Update the dependency-direction guard whenever an internal crate dependency edge changes.
+body: 'When a task adds or removes a direct `orbit-*` dependency in a crate manifest, keep three surfaces aligned in the same change: the consuming crate''s `Cargo.toml`, `ARCHITECTURE.md` dependency diagram/prose, and `scripts/check-dependency-direction.sh`. `make ci-fast` runs the guard script and will reject a manifest-only edge change even when the Rust code compiles.'
+evidence:
+- kind: task
+  reference: ORB-00214
+created_at: 2026-05-21T03:25:58.638337Z
+updated_at: 2026-05-21T03:25:58.638337Z
+created_by: codex
+priority: 120
