+++ crates/orbit-core/Cargo.toml
@@ -29,7 +29,6 @@ serde.workspace = true
 serde_yaml.workspace = true
 sha2.workspace = true
 tempfile = "3"
-tiktoken-rs.workspace = true
 toml.workspace = true
 tracing.workspace = true
