+++ crates/orbit-knowledge/Cargo.toml
@@ -23,6 +23,7 @@ serde_json.workspace = true
 serde_yaml.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
+tiktoken-rs.workspace = true
 toml.workspace = true
 tracing.workspace = true
 tree-sitter = "0.26"
