+++ crates/orbit-knowledge/src/metrics/mod.rs
@@ -1,0 +1,9 @@
+mod ingest;
+#[cfg(test)]
+mod ingest_tests;
+mod summary;
+
+pub use ingest::merge_invocation_trace;
+pub use summary::{
+    DoubleReadSummary, KnowledgeStatsSummary, RatioSummary, TokenInputSummary, aggregate,
+};
